test: expose usb midi stub for pio tests
