+++ firmware/test/USB-MIDI.cpp
@@ -1,0 +1,6 @@
+#include "USB-MIDI.h"
+
+MidiInterfaceStub MIDI;
+USBMidiStub usbMIDI;
+HardwareSerial Serial1;
+
